added simple script
Wrote a simple script that creates a .planter dir in home dir. This is
mostly for testing, as I wanted to make sure scripts run by npx could
interact with the user's file system. Later on, if we want to let users
create aliases, those can be saved in something like
`<home-dir>/.planter/alias.json`, and read during the cli execution.