defined schema for planter objects in the config
Also created a script for generating the default config file, so we can
always easily revert back to an official default state, even if we are
using a variation of the default config JSON for testing.

Also, just gives us a space to extend the config and control how it
should look, if it ever needs to change.