feat: default config is based on Planter class
This allows us to ensure the default config have valid planter data to
read in.

The config scipts use the Planter class as a "schema" to build the
planter objects using the Planter class, and then saves them as JSON in
a file (config.json).

Also updated the schema/Planter class itself, to include a type property
in place of the isDirectory property.